Introduce a generator-compatible decorator and verify stream isolation works correctly with generators across threads.
