+++ tests/python/memory_drain/main.py
@@ -9,31 +9,51 @@ rerun *.rrd
 
 from __future__ import annotations
 
+import queue
 import threading
 import time
-import uuid
+from typing import Any, Iterator
 
 import rerun as rr
 import rerun.blueprint as rrb
 
 
-def job(name: str) -> None:
-    with rr.new_recording("rerun_example_memory_drain", recording_id=uuid.uuid4()):
-        mem = rr.memory_recording()
+@rr.thread_local_stream("rerun_example_memory_drain")
+def job(name: str) -> Iterator[tuple[str, int, bytes]]:
+    mem = rr.memory_recording()
 
-        blueprint = rrb.Blueprint(rrb.TextLogView(name="My Logs", origin="test"))
+    blueprint = rrb.Blueprint(rrb.TextLogView(name="My Logs", origin="test"))
 
-        rr.send_blueprint(blueprint)
+    rr.send_blueprint(blueprint)
 
-        for i in range(5):
-            time.sleep(0.2)
-            rr.log("test", rr.TextLog(f"Job {name} Message {i}"))
+    for i in range(5):
+        time.sleep(0.2)
+        rr.log("test", rr.TextLog(f"Job {name} Message {i}"))
 
-            with open(f"output_{name}_{i}.rrd", "wb") as f:
-                f.write(mem.drain_as_bytes())
+        print(f"YIELD {name} {i}")
+        yield (name, i, mem.drain_as_bytes())
+
+
+def queue_results(generator: Iterator[Any], out_queue: queue.Queue) -> None:
+    for item in generator:
+        out_queue.put(item)
 
 
 if __name__ == "__main__":
-    threading.Thread(target=job, args=("A",)).start()
-    threading.Thread(target=job, args=("B",)).start()
-    threading.Thread(target=job, args=("C",)).start()
+    results_queue: queue.Queue[tuple[str, int, bytes]] = queue.Queue()
+
+    threads = [
+        threading.Thread(target=queue_results, args=(job("A"), results_queue)),
+        threading.Thread(target=queue_results, args=(job("B"), results_queue)),
+        threading.Thread(target=queue_results, args=(job("C"), results_queue)),
+    ]
+    for t in threads:
+        t.start()
+    for t in threads:
+        t.join()
+
+    while not results_queue.empty():
+        name, i, data = results_queue.get()
+
+        with open(f"output_{name}_{i}.rrd", "wb") as f:
+            f.write(data)
